WSTEAM1-623: Move logic back into oEmbed Loader
